Rename normalizer datagen files
